requirements.txt
annotated-types==0.6.0
openai==1.51.1
audioread==3.0.1
blinker==1.7.0
blis==0.7.11
boto3==1.34.139
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
decorator==5.1.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
langchain-aws==0.1.9
langchain-community==0.2.6
langcodes==3.3.0
language-tool-python==2.7.1
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
Markdown==3.5.2
MarkupSafe==2.1.5
msgpack==1.0.8
murmurhash==1.0.10
mutagen==1.47.0
my-voice-analysis==0.7
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
platformdirs==4.2.2
pooch==1.8.1
praat-parselmouth==0.4.3
preshed==3.0.9
protobuf==4.25.3
pyasn1==0.5.1
pycparser==2.22
pydantic==2.6.1
pydantic-core==2.16.2
pydub==0.25.1
pyenchant==3.0.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sklearn==0.0
smart-open==6.4.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.3.0
tqdm==4.66.2
typer==0.9.0
typing-extensions==4.11
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
werkzeug==3.0.1
zipp==3.18.1
httpx==0.27.2
python-dotenv==1.0.1